Remove feature which trows page not found exception when cHash is missing as it is not needed anymore.
It is not needed on TYPO3 higher than 8.5 (or equal), because it is not possible to call Extbase action without cHash (see https://docs.typo3.org/typo3cms/extensions/core/8.7/Changelog/8.5/Breaking-78002-EnforceCHashArgumentForExtbaseActions.html).
It is also not needed to TYPO3 lower than 8.5 because singleview extension provides hashBase customization which use different caching for list view and single views of plugin.